Moved specific layers to corresponding folder
