+++ src/device_p24.py
@@ -1,6 +1,6 @@
 """Provides device class representing a P24"""
 
-from .device import Device, DeviceCapabilities
+from .device import Device, DeviceCapability
 from .utils.connection import Connection
 
 
@@ -8,6 +8,6 @@ class DeviceP24(Device):
     """Device class for a P24 device"""
 
     def __init__(self, conn: Connection):
-        super().__init__(conn, [DeviceCapabilities.GENERAL,
-                                DeviceCapabilities.LOW_LEVEL,
-                                DeviceCapabilities.MID_LEVEL])
+        super().__init__(conn, [DeviceCapability.GENERAL,
+                                DeviceCapability.LOW_LEVEL,
+                                DeviceCapability.MID_LEVEL])
