+++ src/mid_level/layer_mid_level.py
@@ -1,12 +1,12 @@
 """Provices mid level layer"""
 
-from .mid_level.mid_level_current_data import PacketMidLevelGetCurrentData, PacketMidLevelGetCurrentDataAck
-from .mid_level.mid_level_stop import PacketMidLevelStop, PacketMidLevelStopAck
-from .mid_level.mid_level_update import PacketMidLevelUpdate, PacketMidLevelUpdateAck
-from .mid_level.mid_level_types import MidLevelChannelConfiguration
-from .mid_level.mid_level_init import PacketMidLevelInit, PacketMidLevelInitAck
-from .layer import Layer
-from .protocol.protocol import Protocol
+from .mid_level_current_data import PacketMidLevelGetCurrentData, PacketMidLevelGetCurrentDataAck
+from .mid_level_stop import PacketMidLevelStop, PacketMidLevelStopAck
+from .mid_level_update import PacketMidLevelUpdate, PacketMidLevelUpdateAck
+from .mid_level_types import MidLevelChannelConfiguration
+from .mid_level_init import PacketMidLevelInit, PacketMidLevelInitAck
+from ..layer import Layer
+from ..protocol.protocol import Protocol
 
 
 class LayerMidLevel(Layer):
